# openapi.yaml
openapi: 3.0.3
info:
  title: Releasarr API
  version: 0.1.0
  description: |
    Contract derived from the new frontend client. Captures the HTTP interface the
    UI expects today so the backend can be validated and mocked consistently.
  license:
    name: Proprietary
    url: https://github.com/koreetz/releasarr
servers:
  - url: https://api.releasarr.dev/api
    description: Production server
  - url: http://localhost:8001/api
    description: Local development server
security:
  - BearerAuth: []
  - ServiceKeyAuth: []
paths:
...
  /auth/setup:
    get:
      tags: [Auth]
      summary: Whether first-run setup is required
      operationId: getSetupStatus
      security: []
      responses:
        '200':
          description: Whether an admin account still needs to be created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupStatus'
    post:
      tags: [Auth]
      summary: Create the first admin account
      operationId: completeSetup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupPayload'
      responses:
        '201':
          description: The admin account was created and a session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '409':
          description: Setup has already been completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with a username and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginPayload'
      responses:
        '200':
          description: >-
            Access token, plus a refresh cookie (Set-Cookie, httpOnly). The
            cookie's lifetime follows `remember_me`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account is temporarily locked due to repeated failed logins.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange the refresh cookie for a new access token
      operationId: refreshSession
      security: []
      responses:
        '200':
          description: A new access token and a rotated refresh cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Refresh cookie is missing, invalid, or expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: The caller's own profile and effective permissions
      operationId: getMe
      responses:
        '200':
          description: The authenticated user's profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUser'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token obtained from /auth/login or /auth/refresh.
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        A service API key, acting as the user it is bound to. May be combined
        with X-Act-As-User to act as a different user, when the key allows it.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error identifier.
        message:
          type: string
          description: Human-readable explanation of the error.
        details:
          description: Optional key/value data with context for the failure.
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
              - type: array
                items:
                  type: string
...
    UserRole:
      type: string
      enum: [admin, user]
      description: Admin bypasses every per-user permission flag below.
    User:
      type: object
      required:
        - id
        - username
        - role
        - is_active
        - can_view_all_requests
        - can_access_tasks
        - can_access_indexers
        - can_access_logs
        - allowed_root_folders
        - created_at
        - updated_at
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
        can_view_all_requests:
          type: boolean
          description: See every request, not only ones owned by this user.
        can_access_tasks:
          type: boolean
        can_access_indexers:
          type: boolean
        can_access_logs:
          type: boolean
        allowed_root_folders:
          type: array
          items:
            type: string
          description: Library root folders this user may add to. Empty means unrestricted.
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SessionUser:
      allOf:
        - $ref: '#/components/schemas/User'
      description: The authenticated caller's own profile, as returned by GET /auth/me.
...
    SetupStatus:
      type: object
      required: [required]
      properties:
        required:
          type: boolean
          description: Whether first-run setup must run before anything else.
    SetupPayload:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        display_name:
          type: string
          nullable: true
    LoginPayload:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        remember_me:
          type: boolean
          default: false
    LoginResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
        user:
          $ref: '#/components/schemas/SessionUser'